Phase 3e: Extract route handlers to serv/routing/handlers.py
- Extracted complete Route base class (~740 lines) from serv/routes.py
- Moved handler discovery, signature analysis, parameter injection logic
- Includes HTTP method handling, form processing, error handling
- Comprehensive docstrings and examples for Route class usage
- Updated serv/routes.py to comprehensive re-export module for backward compatibility
- Updated routing __init__.py to import from new handlers module
- All 298 tests passing, no functionality changes
- Maintains full backward compatibility through re-exports
