+++ serv/routing/__init__.py
@@ -3,19 +3,14 @@
 # Re-export from existing modules for backward compatibility
 try:
     from serv._routing import Router
-    from serv.routes import Route
+    from serv.routing.handlers import Route
+    from serv.routing.decorators import handle
 
     __all__ = [
         "Router",
         "Route",
+        "handle",
     ]
 except ImportError:
     # If direct imports fail, will be populated later during refactoring
     __all__ = []
-
-# Import handle separately to avoid circular dependency
-try:
-    from serv.routing.decorators import handle
-    __all__.append("handle")
-except ImportError:
-    pass
